Restructuring for testing
