+++ backend/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "author": "",
   "license": "ISC",
@@ -23,6 +23,9 @@
     "uuidv4": "^6.2.13"
   },
   "devDependencies": {
-    "prisma": "^3.12.0"
+    "jest": "^28.1.0",
+    "mock-fs": "^5.1.2",
+    "prisma": "^3.12.0",
+    "supertest": "^6.2.3"
   }
 }
